working analytics/history + minibot connection
