+++ static/gui/static/js/components/BotControl/SetupBot/AddBot.js
@@ -6,8 +6,7 @@ import { faInfoCircle } from '@fortawesome/free-solid-svg-icons';
 library.add(faInfoCircle);
 import { Button } from '../../utils/Util.js';
 import InformationBoxModal from '../../utils/InformationBoxModal.js';
-import RefreshingList from '../GarbageMaybe/RefreshingList.js';
-import BotSearch from './BotSearch.js';
+import RefreshingList from './RefreshingList.js';
 import PortsList from './PortsList.js';
 import {
     INFOBOXTYPE,
@@ -25,18 +24,6 @@ export default class AddBot extends React.Component {
             availableBots: [], // bots connected to Base Station but not GUI
             botList: [],
             power: 50,
-
-            messageVisibility: false, // for flash messages displayed
-            messageType: "",
-            messageContent: "", // what is in the message
-
-            // setup stuff?
-            setupMode: "normal",
-            buttonStatus: false,
-            isSetupReady: false,
-            isSetup: false,
-
-            showPorts: false
         };
 
         // Needed to use a ref for react
@@ -46,7 +33,6 @@ export default class AddBot extends React.Component {
         this.buttonMapListener = this.buttonMapListener.bind(this);
         this.handleArrowKeyDown = this.handleArrowKeyDown.bind(this);
         this.motorPorts = this.motorPorts.bind(this);
-        this.portConfigBttnOnClick = this.portConfigBttnOnClick.bind(this);
     }
 
     portConfigBttnOnClick() {
@@ -61,9 +47,9 @@ export default class AddBot extends React.Component {
     }
 
     componentDidMount() {
-        // setInterval(this.refreshAvailableBots.bind(this), 500);
-        // document.getElementById("setup_control_tab").addEventListener(
-        //     "keydown", this.handleArrowKeyDown);
+        setInterval(this.refreshAvailableBots.bind(this), 500);
+        document.getElementById("setup_control_tab").addEventListener(
+            "keydown", this.handleArrowKeyDown);
     }
 
     /*
@@ -78,9 +64,6 @@ export default class AddBot extends React.Component {
             url: '/active-bots',
         }).then(function (response) {
             _this.state.availableBots = response.data;
-
-            if (response.data.length > 0) _this.state.buttonStatus = false;
-
             // If the Selected Bot (the currently connected bot)
             // is no longer active, remove it from the Selected Bot label (the 
             // currently connected bot label)
@@ -200,6 +183,7 @@ export default class AddBot extends React.Component {
         this.props.setSelectedBotStyle("hidden");
     }
 
+
     getVisionData() {
         const _this = this;
         axios({
@@ -259,68 +243,19 @@ export default class AddBot extends React.Component {
         });
     }
 
-    modeOnChange(value) {
-        if (value == "obj-detection") {
-            this.objectDetectionOnClick();
-            // console.log("obj");
-        } else if (value == "line-follow") {
-            this.lineFollowOnClick();
-            // console.log("line");
-        }
-    }
 
-    flashMessage(value) {
-
-    }
-
-    // finishBotSetup(event) {
-    //     // let _this = this;
-    //     // TODO: run all commands for bot setup here
-    //     this.addBotListener(event);
-    //     // if (_this.state.setupMode == "obj-detection") {
-    //     //     _this.objectDetectionOnClick();
-    //     //     // console.log("obj");
-    //     // } else if (_this.state.setupMode == "line-follow") {
-    //     //     _this.lineFollowOnClick();
-    //     //     // console.log("line");
-    //     // }
-    // }
 
     render() {
         const _this = this;
         return (
-            <div id="bot-setup-area">
-                {/* // Set tabindex to -1 so that this div is in focus to caputure 
+            // Set tabindex to -1 so that this div is in focus to caputure 
             // the keyboard event handler for arrow key movement
-            // <div id="setup_control_tab" tabIndex="-1" className="container-fluid control"> */}
-                <div id="bot-setup" className="control-option">
-                    {/* <div className="row">
-                        <div className="col d-flex">
-                            <h3 className="small-title"> Setup the Bot <span className="info-icon"><FontAwesomeIcon icon='info-circle' /></span></h3>
-                            <button className="btn btn-secondary ml-auto">Search for bots</button>
-                        </div>
-                    </div>
-                    <div className="form-group row">
-                        <label className="col-sm-4 col-form-label">Select Bot:</label>
-                        <div className="col-sm-8">
-                            <RefreshingList ref={this.refreshingBotListRef} />
-                        </div>
-                    </div> */}
-                    <BotSearch ref={this.refreshingBotListRef} />
-                    <div className="form-group row">
-                        <label className="col-sm-4 col-form-label">Select Mode:</label>
-                        <div className="col-sm-8">
-                            <select className="custom-select custom-select-sm" onChange={(event) => this.state.setupMode = event.target.value}>
-                                <option value="">Normal</option>
-                                <option value="obj-detection">Object Detection</option>
-                                <option value="line-follow">Line Follow</option>
-                            </select>
-                        </div>
-                    </div>
-                    {/* <div className="row">
+            <div id="setup_control_tab" tabindex="-1" className="container-fluid control">
+                <div className="container">
+                    <div className="row">
                         <div className="col text-center">
                             <br />
-                            <h3 className="small-title"> Minibot Setup </h3>
+                            <p className="small-title"> Minibot Setup </p>
                         </div>
                     </div>
                     <div className="row">
@@ -329,7 +264,7 @@ export default class AddBot extends React.Component {
                                 <label className="white-label"> Available Bots: &nbsp; </label>
                                 <RefreshingList ref={this.refreshingBotListRef} />
                             </div>
-                            <Button id="add-bot" name="Add Bot" onClick={this.addBotListener} />
+                            <button className="btn btn-secondary" onClick={this.addBotListener}>Add Bot</button>
                         </div>
                     </div>
                     <div className="row">
@@ -337,29 +272,31 @@ export default class AddBot extends React.Component {
                             <div className="element-wrapper">
                                 <label id="selected-bot" style={this.props.selectedBotStyle}>
                                     Connected to: &nbsp; &nbsp;
-                                <span id="botName">
+                                    <span id="botName">
                                         {_this.getSelectedBotText()}
                                     </span>
                                 </label>
                             </div>
-                            <Button id="remove-bot" name="Remove"
-                                onClick={() => _this.deleteBotListener()}
-                                style={_this.props.selectedBotStyle} />
-                        </div>
-                    </div> */}
-                </div>
-                <div className="control-option">
-                    <div className="row">
-                        <div className="col d-flex justify-content-end">
-                            <button className="btn btn-secondary" disabled={this.state.buttonStatus} onClick={this.addBotListener}>Finish Bot Setup</button>
+                            <button className="btn btn-secondary" style={_this.props.selectedBotStyle} onClick={() => _this.deleteBotListener()}>Remove</button>
                         </div>
                     </div>
                 </div>
-                <br /><br />
+                <div className="container">
+                    <div className="row">
+                        <div className="col horizontalDivCenter">
+                            <p className="small-title"> Custom Modes </p>
+                            <button className="btn btn-success element-wrapper mr-1" onClick={() => this.objectDetectionOnClick()}>Object Detection</button>
+                            <button className="btn btn-primary element-wrapper mr-1" onClick={() => this.lineFollowOnClick()}>Line Follow</button>
+                        </div>
+                    </div>
+                    <br />
+                </div>
+                <br />
+                <br />
+                <br />
                 <div id="port-config" className="control-option">
                     <div className="row">
                         <div className="col">
-
                             <h3 className="small-title">
                                 <input id="portConfigBttn"
                                     width="25" height="25" type="image"
@@ -388,59 +325,8 @@ export default class AddBot extends React.Component {
                     </div>
                     <InformationBoxModal type={INFOBOXTYPE.PORT} />
                 </div>
-                {/* <div className="control-option">
-                    <div className="row">
-                        <div className="col horizontalDivCenter">
-                            <p className="small-title">Movement </p>
-                        </div>
-                    </div>
-                    <div className="row">
-                        <div className="col text-center">
-                            <button className="button-movement" onClick={() => this.buttonMapListener("forward")}>forward</button>
-                            <br />
-                            <button className="button-movement" onClick={() => this.buttonMapListener("left")}>left</button>
-                            <button className="button-stop" onClick={() => this.buttonMapListener("stop")}>stop</button>
-                            <button className="button-movement" onClick={() => this.buttonMapListener("right")}>right</button>
-                            <br />
-                            <button className="button-movement" onClick={() => this.buttonMapListener("backward")}>backward</button>
-                            <br />
-                            <br />
-                            <form className="horizontalDivCenter">
-                                <div>
-                                    <label className="white-label"> Power: &nbsp; </label>
-                                </div>
-                                <input id="custom-range-1" className="custom-range" name="wheel_power" type="range" min="0" max="100"
-                                    step="5" onChange={evt => this.updatePowerValue(evt)} />
-                            </form>
-                        </div>
-                    </div>
-                </div>
-                <div className="control-option">
-                    <div className="row">
-                        <div className="col horizontalDivCenter">
-                            <p className="small-title"> Speech Recognition </p>
-                        </div>
-                    </div>
-                    <div className="col horizontalDivCenter">
-                        <SpeechRecognition selectedBotName={this.props.selectedBotName}
-                            float="right" />
-                    </div>
-                    <div className="col horizontalDivCenter">
-                        <label id="speech_recognition_feedback_box" />
-                    </div>
-                </div> */}
 
-                {/* <div className="control-option">
-                    <div className="row">
-                        <div className="col horizontalDivCenter">
-                            <p className="small-title"> Custom Modes </p>
-                            <button className="btn btn-success element-wrapper mr-1" onClick={() => this.objectDetectionOnClick()}>Object Detection</button>
-                            <button className="btn btn-primary element-wrapper mr-1" onClick={() => this.lineFollowOnClick()}>Line Follow</button>
-                        </div>
-                    </div>
-                </div> */}
-                {/* </div > */}
-
+                <br />
             </div >
         );
     }
